requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
python-dateutil
requests
youtube_dl
pynacl